use local page jsx for contentful page if exists
Sadly, it seems you can only pass a context to pages in src/templates from createPage. A contentful page jsx has to hardcode it's own slug on the page filter because of this.
I also considered passing the page component path to the pageTemplate via context, but to render the page component, the template would have to dynamically import it, which adds too much complexity imo.