Add mobile navigation (#67)
* Add mobile navigation

* Add Responsive rendering of NavBars

* Add CSSTransition to MobileNavBar

* Make NavBar fixed to top and make content scroll instead of window

* Change MediaQuery and styling of NavBar

* Refactor renderfunctions in MobileNavBar

* Fix build

The build is broken, because we can't use <MediaQuery> without giving it a values object, because we're rendering server side. It seems to be needed only once, but maybe we should come up with a more detailed default with other media properties. The Devices.jsx on the dry-media-queries branch seems like a good place.
See [the react-responsive-docs](https://github.com/contra/react-responsive#server-rendering) and [this issue](https://github.com/gatsbyjs/gatsby/issues/2142)